+++ tests/integration/cli/test_explain.py
@@ -30,7 +30,7 @@ class ExplainIntegrationScenarios(TestCase):
 
         with TestDir() as project_dir:
             # Create project
-            run(self, "create", "-o", project_dir)
+            run(self, "project", "create", "-o", project_dir)
 
             model_name = "my-model"